requirements.txt
python-dotenv
mysqlclient
SQLAlchemy
streamlit
#mariadb
requests
beautifulsoup4
pandas
pdfkit
Jinja2